backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
PyPDF2>=3.0.1
python-multipart>=0.0.9
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
httpx>=0.27.0
duckduckgo-search>=7.0.0
ddgs>=9.0.0
python-docx>=1.0.0
python-pptx>=1.0.0
